+++ src/winml/modelkit/analyze/core/runtime_checker_query.py
@@ -25,6 +25,7 @@ from ...onnx import (
     infer_onnx_shapes,
     remove_optional_from_type_annotation,
 )
+from ...onnx.external_data import try_load_external_initializer_array
 from ...pattern.base import (
     get_pattern_input_generator,
     get_registered_pattern_input_generators,
@@ -629,103 +630,6 @@ def _normalize_type_var_annotation(type_value: str) -> str:
         return SupportedONNXType.from_onnx_type(type_value).annotation
     except ValueError:
         return SupportedONNXType.normalize_annotation(type_value)
-
-
-# Query conditions are later normalized with make_hashable, which materializes
-# numpy arrays. Keep external initializer loading bounded until the broader
-# condition pipeline can stay lazy end-to-end.
-MAX_EXTERNAL_INITIALIZER_BYTES_FOR_QUERY = 1024 * 1024
-
-
-def _get_external_tensor_info(tensor: onnx.TensorProto) -> tuple[str | None, int, int | None]:
-    """Extract location, offset, and length metadata for an external tensor."""
-    info = {entry.key: entry.value for entry in tensor.external_data}
-    location = info.get("location")
-    offset = int(info.get("offset", "0"))
-    length = int(info["length"]) if "length" in info else None
-    return location, offset, length
-
-
-def _tensor_proto_dtype_to_np_dtype(tensor_type: int) -> np.dtype[Any]:
-    """Convert a TensorProto dtype enum to a numpy dtype."""
-    try:
-        from onnx.helper import tensor_dtype_to_np_dtype as onnx_tensor_dtype_to_np_dtype
-    except ImportError:
-        from onnx.mapping import TENSOR_TYPE_TO_NP_TYPE
-
-        return np.dtype(TENSOR_TYPE_TO_NP_TYPE[tensor_type])
-
-    return np.dtype(onnx_tensor_dtype_to_np_dtype(tensor_type))
-
-
-def try_load_external_initializer_array(
-    tensor: onnx.TensorProto,
-    model_path: str | Path | None,
-) -> np.ndarray | None:
-    """Load a small external initializer from disk for runtime-query conditions.
-
-    Returns None when the model path is unavailable, metadata is incomplete,
-    the sidecar file is missing, or the tensor is too large for the current
-    hash-based condition pipeline.
-    """
-    if tensor.data_location != onnx.TensorProto.EXTERNAL or model_path is None:
-        return None
-
-    location, offset, length = _get_external_tensor_info(tensor)
-    if location is None:
-        return None
-
-    model_path = Path(model_path)
-    data_path = Path(location)
-    if not data_path.is_absolute():
-        data_path = model_path.parent / data_path
-    if not data_path.exists():
-        return None
-
-    try:
-        np_dtype = _tensor_proto_dtype_to_np_dtype(tensor.data_type)
-        shape = tuple(tensor.dims)
-        numel = int(np.prod(shape)) if shape else 1
-        expected_bytes = numel * np_dtype.itemsize
-        if length is not None and length < expected_bytes:
-            logger.debug(
-                "External initializer %s length %s is smaller than expected %s",
-                tensor.name,
-                length,
-                expected_bytes,
-            )
-            return None
-        if expected_bytes > MAX_EXTERNAL_INITIALIZER_BYTES_FOR_QUERY:
-            logger.debug(
-                "Skipping external initializer %s for query conditions because %s bytes exceeds %s",
-                tensor.name,
-                expected_bytes,
-                MAX_EXTERNAL_INITIALIZER_BYTES_FOR_QUERY,
-            )
-            return None
-
-        with data_path.open("rb") as f:
-            f.seek(offset)
-            arr = np.fromfile(f, dtype=np_dtype, count=numel)
-
-        if arr.size != numel:
-            logger.debug(
-                "External initializer %s only yielded %s elements, expected %s",
-                tensor.name,
-                arr.size,
-                numel,
-            )
-            return None
-
-        return arr.reshape(shape)
-    except Exception as e:
-        logger.debug(
-            "Failed to read external initializer %s from %s: %s",
-            tensor.name,
-            data_path,
-            e,
-        )
-        return None
 
 
 def _get_pattern_type_var_conditions(
